well, we are only deploying to sbcl
so make sure that travis doesn't spend time spinning up the other contexts. This
whole codebase is portable to at least ccl, ecl, and abcl except for the call to
join-threads in the #'-main CLI context. Bordeaux Threads is not doing the right
thing, so I'll save debugging that for another day.